Redoing the way fake file system.
Major changes to the zipped files:
* All the files and directories are stored on first read.
* It is still the case that only one level of embedded zip files are unzipped
* Changed what is associated with files to be the index into zip file, along with a subindex for embedded zipped files.
* The stored files are use "/" for the path separator

For the normal case, added a special case/hack so that "zh/tw" is added to the directories returned by `find_all_dirs_shim`. Needed because the zip files are not looked into.

There is a bug in `find_all_dirs_shim`: it returns "zh" (both for zipped/no-zip cases) even though there aren't really any rules for 'zh'. Changes to prefs.rs allow it to find zh-tw, but that's not really right.