Fixes #110 -- navigation inside of leaves
Propogated NavNodeOffset to all the places it needed to be.

I needed to add code to `find_match` to track the match while navigating.

There are also some safeguards in case the same ID is used. That's a different bug.

For some reason, pausing changed slightly for two tests, so those were fixed.