Replace unimplemented! with todo! in ReadTo command
`todo!` more accurately describes the situation.

I don't know of any software that uses these interfaces; I think all of them just pass the keystroke on through. In this case, it is cntrl+shift+start/end and that results in "Error in Navigating Math" because it isn't implemented in navigation.yaml. At the time (a few years ago), I couldn't figure out how to implement this. Perhaps with more experience, I can figure it out now. No one has asked for it though.